NPMCLI-75 fix race conditions
This fixes mutliple race conditions that resulted in flaky tests.

Its also a problem for very slow file systems or in cases
where the CLI is used programatically.

The async "branches" weren't using callbacks or another way
to resume the execution, so it was a bet on them finishing fast
enough.

That meant in practice that when the CLI was finished, the created
files usually weren't available yet.