NPMCLI-69 Set up tab completion for bash, zsh and PowerShell
A new script, run on npm postinstall or through a command, generates a shell script for
automatic tab completion in bash, zsh and PowerShell, then copies them to an appropriate
locations based on current permissions. If copied to the user home directory,
init scripts (like .bashrc) are extended with a snippet loading said file on
shell initialization.

Analogously, a preuninstall npm hook (or uninstall command) removes those files wherever possible with
current permissions. Additionally, resulting empty directories (like ~/.mgnl) are removed too.

~ Extract separate module for file snippet management

FileSnippetManager handles parts of files, like insertion and deletion, without
affecting the rest of that file. For example, this is useful when inserting
sourcing code to .bashrc for tab autocompletion, but wanting to prevent duplicates
and be able to remove it on uninstall.

~ Details on tab autocompletion for PowerShell

- Autocompletion is achieved through a custom function named `mgnl` using
  ValidateSet on parameters
- A check if the command `powershell` is available on the current system
  determines whether PS is installed

~ Extend tab completion with parameters (bash and zsh only)

- Add 'options' property to objects provided from commands.js
- List those options in the autocompletion bash script if command
  prompt line looks like expecting a parameter flag
- Fall back to default completion if neither command nor parameter
  flag may be expected