NPMCLI-115 ensure diff hunks start operations from same line
we're using a diff library and manipulate the diff file itself to put back empty lines and comments after adding new component availability.
In some cases this seems to produce a wrong diff file with hunk operations starting from different lines which end up screwing the file the patch
is applied to.